ci: Fix workflow
The changes made to the docs.yml involve adjustments in the workflow to now only trigger on push to the main branch rather than the main and catalog branches. The naming has been adjusted from 'Run npm build' to 'Build docs' and the command to run has changed from 'pnpm build' to 'pnpm build:build'. These changes should allow for a more readable workflow and improve the process of building and publishing the documentation.