ci: Renamed GitHub workflow directory
The workflow directory in the .github folder was incorrectly named as 'workflow'. This typo has been corrected by renaming it to 'workflows'. Additionally, several adjustments have been made inside the docs.yml file to trigger build only on push to the main branch, enhancing readability and efficiency in the documentation build and publication process.