chore: 🙈 Update .gitignore to include 'docs' folder
The 'docs' folder has been added to the .gitignore file. This means that files and directories under 'docs' will be ignored by Git and will not be tracked or reported as untracked. This change is useful for excluding documentation files that may be updated frequently but do not need to be version-tracked.